decimal 问题

+++ oexpress_dw/oexpress_source_binlog_h_his.py
@@ -352,7 +352,7 @@ def run_sqoop_check_table(mysql_db_name, mysql_table_name, conn_id, hive_table_n
                 #     data_type = result[1] + "(" + str(result[2]) + "," + str(result[3]) + ")"
                 # 有json表读取insert 部分，此处切换为double
             elif result[1] == 'decimal':
-                data_type = 'double'
+                data_type = 'decimal'
             elif result[1] == 'mediumint':
                 data_type = 'int'
             else:
